+++ pandas/core/base.py
@@ -1210,7 +1210,7 @@ class IndexOpsMixin(OpsMixin):
         3
 
         >>> ser.searchsorted([0, 4])
-        array([0, 3],  dtype=int64)
+        array([0, 3])
 
         >>> ser.searchsorted([1, 3], side='left')
         array([0, 2], dtype=int64)
